add responsive card
